Moving the content to md files
The following template html files are now written in markdown:
c4p, ciudad, codigo_conducta, organizacion, patrocinios, and viajes.

Special HTML code is added for patrocinios,
where the style of the buttons, tables, and cards was not simple
to add to a markdown file.
Organizacion includes a file that generates que list of organization
team.

The page for previous event was left untouched due to the cards
that can be found in the HTML file.